Docker mount /var/cache,apt/lib as volumes
Use cache shared between packer builds.
Also add sync for all scripts.

+++ script/minimize.sh
@@ -14,7 +14,7 @@ echo "==> Removing documentation"
 dpkg --list | awk '{ print $2 }' | grep -- '-doc$' | xargs apt-get -y purge
 echo "==> Removing development tools"
 #dpkg --list | grep -i compiler | awk '{ print $2 }' | xargs apt-get -y purge
-#apt-get -y purge cpp gcc g++ 
+#apt-get -y purge cpp gcc g++
 apt-get -y purge build-essential
 echo "==> Removing default system Ruby"
 apt-get -y purge ruby ri doc
@@ -27,10 +27,14 @@ apt-get -y purge ppp pppconfig pppoeconf
 echo "==> Removing other oddities"
 apt-get -y purge popularity-contest installation-report landscape-common wireless-tools wpasupplicant ubuntu-serverguide
 
-# Clean up the apt cache
-apt-get -y autoremove --purge
-apt-get -y autoclean
-apt-get -y clean
+# NOTE(bogdando) Don't do reboot/sleep when used from docker templates.
+# only atlas templates define the VERSION.
+if [ "${VERSION}" ]; then
+  # Clean up the apt cache
+  apt-get -y autoremove --purge
+  apt-get -y autoclean
+  apt-get -y clean
+fi
 
 # Clean up orphaned packages with deborphan
 apt-get -y install deborphan
@@ -41,11 +45,14 @@ apt-get -y purge deborphan dialog
 
 echo "==> Removing man pages"
 rm -rf /usr/share/man/*
-echo "==> Removing APT files"
-find /var/lib/apt -type f | xargs rm -f
 echo "==> Removing anything in /usr/src"
 rm -rf /usr/src/*
 echo "==> Removing any docs"
 rm -rf /usr/share/doc/*
-echo "==> Removing caches"
-find /var/cache -type f -exec rm -rf {} \;
+if [ "${VERSION}" ]; then
+  echo "==> Removing caches"
+  find /var/cache -type f -exec rm -rf {} \;
+  echo "==> Removing APT files"
+  find /var/lib/apt -type f | xargs rm -f
+fi
+sync
