+++ script/update.sh
@@ -8,6 +8,11 @@ if [[ $UPDATE  =~ true || $UPDATE =~ 1 || $UPDATE =~ yes ]]; then
     apt-get -y install sudo
     echo "==> Performing dist-upgrade (all packages and kernel)"
     apt-get -y dist-upgrade --force-yes
-    reboot
-    sleep 160
+    sync
+    # NOTE(bogdando) Don't do reboot/sleep when used from docker templates.
+    # only atlas templates define the VERSION.
+    if [ "${VERSION}" ]; then
+      reboot
+      sleep 160
+    fi
 fi
